requirements.txt
# Web Framework
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0,<2.0.0

# Redis & Caching
redis>=5.0.0
hiredis>=2.0.0

# Task Queue
celery>=5.3.0,<6.0.0
flower>=2.0.0,<3.0.0  # Celery monitoring UI

# ML & NLP
transformers>=4.36.0,<5.0.0
torch>=2.1.0,<3.0.0
sentence-transformers>=2.2.0,<3.0.0
faiss-cpu>=1.7.0,<2.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0

# Data Processing
pandas>=2.1.0,<3.0.0
feedparser>=6.0.0
newspaper4k>=0.9.3.1
beautifulsoup4>=4.12.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0  # JWT tokens
passlib[bcrypt]>=1.7.0  # Password hashing
python-multipart>=0.0.6  # File uploads

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0  # For testing async endpoints
black>=23.0.0  # Code formatting
flake8>=7.0.0  # Linting
mypy>=1.8.0  # Type checking
ipython>=8.19.0  # Better REPL
ipdb>=0.13.0  # Debugging